requirements.txt
alpaca-py==0.43.2
websockets==16.0
pandas==2.2.3
numpy==2.2.6
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
apscheduler==3.11.2
pyyaml==6.0.3
python-dotenv==1.2.1
aiohttp==3.13.3
rich==14.3.3
jinja2==3.1.6
pytest==8.4.2
pytest-asyncio==0.23.8
pytest-mock==3.15.1
httpx==0.28.1
requests==2.32.5
lxml==5.4.0
pytz==2025.2
